Fixed one more bug for nat-piercer protocol.
Problem was that after the relay user left the channel, and a new one was promoted, there the protocol would stop, nothing more would happen. I fixed it by giving the new relay the list of connected peers, so he knows who is connected to him in the voice channel, and from there he can start keepalive thread to send heartbeat messages to the other users and keep connections alive. This stage of the protocol works great now. One more fix i did: the HOLE-PUNCHING thread would send punches for only 5 seconds from the user that wants to connect to the relay user. I changed that, now it sends punches for an undetermined period, until connection is established.