Users can now send messages to each other.
Added a new thread that waits for keyboard input and saves the input as a message. Messages are saved in a payload, that will represent data packets (for now simple strings to test that this works). In the main loop, when receiving a DATA... starting message, it is handled by the relay, he will send the DATA packet to all other peers connected to him. For the moment i keep star topology even for message sending. Might change to star topology in the future.